+++ RELEASE_NOTES.md
@@ -31,8 +31,8 @@ Prolog adapter and integration files
 * CHANGED: Moved all common setup code from the Bash shell integration scripts
 to a shared setup script.
 
-* CHANGED: Moved all common setup code from the PowerShell shell integration
-scripts to a module.
+* CHANGED: Moved all common setup code from the PowerShell integration scripts
+to a module.
 
 * FIXED: The PowerShell integration scripts were not setting the `LOGTALKUSER`
 environment variable correctly when not defined by the user.
@@ -61,6 +61,12 @@ probability distributions.
 
 Tools
 -----
+
+* CHANGED: Moved all common setup code from the `lgtdoc` tool Bash shell
+scripts to a shared setup script.
+
+* CHANGED: Moved all common setup code from the `lgtdoc` tool PowerShell
+integration scripts to a module.
 
 * ADDED: Predicates `write_max_depth/1` and `set_write_max_depth/1` to the
 `debugger` tool.